An attemp at implementing IDocHostUIHandlerDispatch. Does not compile

+++ GINA/global.h
@@ -51,7 +51,7 @@ EXTERN size_t gUsername_len;
 EXTERN const wchar_t gEncryptedTag[];
 EXTERN const size_t gEncryptedTag_len;
 EXTERN wchar_t gEncryptedRandomSelfservePassword[];
-EXTERN const DWORD gEncryptedRandomSelfservePassword_len;
+EXTERN const size_t gEncryptedRandomSelfservePassword_len;
 
 #define AUCUN_PWLEN (16)
